requirements.txt
livekit-agents>=0.8.0
livekit-plugins-google>=0.4.0
python-dotenv>=1.0.0
psutil>=6.0.0
pyautogui>=0.9.54
pyperclip>=1.9.0
screeninfo>=0.8.1
Pillow>=10.0.0
requests>=2.31.0
